flask-app/requirements.txt
flask
folium
flask_sqlalchemy
mysqlclient
python-dotenv
waitress
pandas
mysql-connector-python